refactor(prompt_cache): move PromptCacheStore to dedicated module (#365)
